requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
chromadb
pdfplumber
pandas
numpy
PyYAML
openai
